Project settings: GitForm
Create a dedicated git project settings form component.
Move the option to force-push to that new form
